Fix for persisting Werkaanbiedingen from Leerling Fix for Error 400 Bad request and 500 internal server error when persisting

+++ app/src/main/java/com/talentcoach/id11/id11_android/repositories/LeerlingRepository.kt
@@ -11,15 +11,18 @@ import java.net.URL
 class LeerlingRepository {
     val url = "http://projecten3studserver11.westeurope.cloudapp.azure.com/api/leerlingen/"
 
-    fun getLeerlingById(id: Long): Leerling {
+    fun getLeerlingById(id: Int): Leerling {
         val leerling: Leerling?
 
         try {
             val json = URL(url + id).readText()
+
             leerling = Klaxon().parse<Leerling>(json)
+
         } catch (e: Exception) {
             throw e
         }
+
         return leerling!!
     }
